Move `BlackBox{Select,Prepare}` next to `{Select,Prepare}Oracle`
