+++ qualtran/bloqs/block_encoding/__init__.py
@@ -19,8 +19,6 @@ from qualtran.bloqs.block_encoding.chebyshev_polynomial import (
     ScaledChebyshevPolynomial,
 )
 from qualtran.bloqs.block_encoding.lcu_block_encoding import (
-    BlackBoxPrepare,
-    BlackBoxSelect,
     LCUBlockEncoding,
     LCUBlockEncodingZeroState,
 )
